Fix up the last bugs

+++ src/commands/remove.rs
@@ -1,5 +1,7 @@
 use std::collections::{HashMap, HashSet};
 
+use data::Topic;
+
 use crate::unit::loader::UnitId;
 
 use super::{Module, Status};
@@ -8,6 +10,8 @@ use super::{Module, Status};
 pub struct Remove {
     #[clap(long)]
     dry_run: bool,
+    #[clap(long)]
+    no_clean: bool,
     topics: Option<Vec<String>>,
 }
 
@@ -19,26 +23,44 @@ impl Remove {
         ctx: &mut crate::context::Context,
     ) -> super::CmdResult {
         println!("Removing units:");
-        if let Some(topics) = self.topics.as_ref() {
-            let topics = HashSet::from_iter(topics.iter().map(String::as_str));
-            filter_modules(topics, root, &mut modules);
-        }
 
-        // Report bad user query
-        if !modules.values().any(|m| m.status == Status::Ready) {
-            if let Some(topics) = &self.topics {
-                super::no_units_match(topics)
+        let mut errn = 0;
+
+        let topics = self.topics.as_ref().map(|topics| {
+            let (topics, errs) = super::register_topics(topics, ctx);
+            errn += errs;
+            topics
+        });
+
+        if errn == 0 {
+            if let Some(topics) = topics.as_ref() {
+                filter_modules(topics, root, &mut modules);
             }
-        }
 
-        let (removed, errn) = super::remove_modules(root, &mut modules, self.dry_run, false, ctx);
+            // Report bad user query
+            if !modules.values().any(|m| m.status == Status::Ready) {
+                if let Some(topics) = &self.topics {
+                    super::no_units_match(topics)
+                }
+            }
 
-        for package in removed
-            .into_iter()
-            .flat_map(|(_id, package)| package.completed)
-        {
-            if let Err(err) = ctx.database().removed(&package) {
-                tracing::error!("Failed to update package database: {err:#}");
+            let (removed, errs) =
+                super::remove_modules(root, &mut modules, self.dry_run, false, ctx);
+            errn += errs;
+
+            for package in removed
+                .into_iter()
+                .flat_map(|(_id, package)| package.completed)
+            {
+                if let Err(err) = ctx.database().removed(&package) {
+                    tracing::error!("Failed to update package database: {err:#}");
+                }
+            }
+
+            // Clean old deployment
+            if !self.no_clean {
+                println!("\n\nCleaning up old packages:");
+                super::clean_deployed(vec![], topics, ctx);
             }
         }
 
@@ -64,7 +86,7 @@ impl Remove {
 
 /// Filter modules for removal based on a set of topics. Member units depend on
 /// their parents, so we need to propagate removal down the tree.
-fn filter_modules(topics: HashSet<&str>, root: UnitId, modules: &mut HashMap<UnitId, Module>) {
+fn filter_modules(topics: &HashSet<Topic>, root: UnitId, modules: &mut HashMap<UnitId, Module>) {
     let mut stack = vec![(vec![root], false)];
     while let Some((members, remove_parent)) = stack.last_mut() {
         match members.pop() {
@@ -73,7 +95,7 @@ fn filter_modules(topics: HashSet<&str>, root: UnitId, modules: &mut HashMap<Uni
                 let remove = *remove_parent
                     || match &module.unit.as_ref().unwrap().topic {
                         // Remove if topics match
-                        Some(topic) => topics.contains(topic.name()),
+                        Some(topic) => topics.contains(&topic),
                         // Keep otherwise
                         None => false,
                     };
